Added support for Runner
I was trying to add support for Runner, which requires the following:
- Keeping a .sketchplugin file in the root of the repo so the plugin
is available in the `install` list.
- Adding icons and descriptions for the commands available in `run`
- Adding icon.png

To do so I tried the following:

- Removing build folder from .gitignore

- Created .eslintignore and added the build folder, so it doesn't raise
errors when triggering the pre-commit hook.

- Added icons & description fields to manifest.json.

- Added icon.png

- Manually copied the icon.png from the src to the build folder. Ideally
skpm could take care of copying those assets to the Resources folder.

An extra thought: If skpm adds those Runner-specific fields and asset we
could even prepare a icon.sketch template in the src folder, that would
then be automagically exported as a png in the Resources folder.

Something else to figure out: In development mode Runner doesn't pick up
icons or descriptions. Probably something to do with the symlink folder?